LayoutTests/fast/xmlhttprequest/xmlhttprequest-multiple-sync-xhr-during-unload.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/js-test.js"></script>
<iframe id="testFrame" src="resources/xmlhttprequest-multiple-sync-xhr-during-unload-iframe.html"></iframe>
<script>
description("Makes sure that we throw when doing sync XHRs during unload.");
jsTestIsAsync = true;

frameDidUnload = (_xhrExceptionCount) => {
    xhrExceptionCount = _xhrExceptionCount;
    shouldBe("xhrExceptionCount", "5");
    finishJSTest();
}

onload = () => {
    testFrame.remove();
}
</script>
</html>
